requirements.txt
streamlit
python-dotenv
pydantic
langchain
langchain-core
langchain-openai
langchain-community
